LPD-27031 LPD-27034 portal-search-tuning-rankings-web-test: Add BaseRankingUpgradeProcessTestCase class
https://liferay.atlassian.net/browse/LPD-27034
